+++ judge-impl/src/main/java/com/dwarfeng/judge/impl/handler/JobLocalCacheHandlerImpl.java
@@ -18,10 +18,7 @@ import com.dwarfeng.subgrade.stack.exception.HandlerException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 @Component
 public class JobLocalCacheHandlerImpl implements JobLocalCacheHandler {
@@ -92,7 +89,11 @@ public class JobLocalCacheHandlerImpl implements JobLocalCacheHandler {
                 transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class
         )
         public boolean exists(LongIdKey key) throws Exception {
-            return sectionMaintainService.exists(key);
+            Section section = sectionMaintainService.getIfExists(key);
+            if (Objects.isNull(section)) {
+                return false;
+            }
+            return section.isEnabled();
         }
 
         @Override